requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose
passlib[bcrypt]
alembic
httpx
psycopg2-binary
pydantic[email]
python-multipart